requirements.txt
numpy==1.21.2
pandas==1.3.4
pretty-midi==0.2.9
pytorch-lightning==1.5.2
scikit-learn==1.0.1
scipy==1.7.1
torch==1.9.1
torchtext==0.10.1
transformers==4.10.2
setuptools==59.5.0
hydra-core==1.3.2
omegaconf==2.3.0
